// FrontEnd/Cervejaria.FrontEnd/ClientApp/src/app/Models/ClienteFornecedor.ts
export interface ClienteFornecedor {
    id: number | null;
    nome: string | null;
    cnpjCpf: string | null;
    ie: string | null;
    tipoUsuario: number | null;
    telefone: string | null;
    contato1: string | null;
    email: string | null;
    cep: string | null;
    numero: number | null;
    complemento: string | null;
}